데몬 어택 등등 추가
데몬 어택 등등 추가

+++ GameFramework/Player.cpp
@@ -41,13 +41,13 @@ void Player::createPlayer()
 }
 
 //-------------------------------------------
-void Player::setHP(int hp)
+void Player::setHP(float hp)
 {
 	mPlayerHP += hp;
 }
 
 //-------------------------------------------
-int Player::getHP()
+float Player::getHP()
 {
 	return mPlayerHP;
 }
